+++ automerge-backend/src/backend.rs
@@ -4,10 +4,10 @@ use crate::op_handle::OpHandle;
 use crate::actor_map::ActorMap;
 use crate::op_set::OpSet;
 use crate::ordered_set::{OrdDelta, OrderedSet};
-use crate::patch::{Diff, Patch, PendingDiff};
+use crate::pending_diff::PendingDiff;
 use crate::undo_operation::UndoOperation;
 use crate::time;
-use automerge_protocol::{ActorID, ChangeHash, ObjType, OpID, ObjectID, Key, ReqOpType, RequestKey, OpRequest, OpType, Value, Operation};
+use automerge_protocol::{ActorID, ChangeHash, ObjType, OpID, ObjectID, Key, ReqOpType, RequestKey, OpRequest, OpType, Value, Operation, Diff, Patch};
 use crate::{Change, ChangeRequest, ChangeRequestType};
 use std::borrow::BorrowMut;
 use std::cmp::max;
