+++ automerge-backend/src/lib.rs
@@ -23,14 +23,12 @@ mod object_store;
 mod op_handle;
 mod op_set;
 mod ordered_set;
-mod patch;
+mod pending_diff;
 mod undo_operation;
 mod serialize;
 mod time;
 mod actor_map;
 
-pub use automerge_protocol::{ActorID, ChangeHash, ObjType, ObjectID, ElementID, Key, DataType, Value, OpType, Operation, Change, ChangeRequest, ChangeRequestType,};
-
-pub use crate::patch::{DiffEdit, Patch, Diff, MapDiff, SeqDiff, ObjDiff};
+pub use automerge_protocol::{ActorID, ChangeHash, ObjType, ObjectID, ElementID, Key, DataType, Value, OpType, Operation, Change, ChangeRequest, ChangeRequestType, DiffEdit, Patch, Diff, MapDiff, SeqDiff, ObjDiff};
 pub use backend::Backend;
 pub use error::AutomergeError;
