+++ automerge-protocol/src/utility_impls/mod.rs
@@ -3,3 +3,4 @@ mod object_id;
 mod key;
 mod element_id;
 mod value;
+mod diff;
